Ported to new test images
