+++ test/conformance/helpers/channel_tracing_test_helper.go
@@ -17,19 +17,14 @@ limitations under the License.
 package helpers
 
 import (
-	"context"
 	"fmt"
-	"net/http"
 	"testing"
-	"time"
 
-	ce2 "github.com/cloudevents/sdk-go/v2"
+	cloudevents "github.com/cloudevents/sdk-go/v2"
 	cetest "github.com/cloudevents/sdk-go/v2/test"
 	"github.com/google/uuid"
 	"github.com/openzipkin/zipkin-go/model"
-	"go.opentelemetry.io/otel/api/trace"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"knative.dev/pkg/test/zipkin"
 
 	tracinghelper "knative.dev/eventing/test/conformance/helpers/tracing"
 	"knative.dev/eventing/test/lib"
@@ -37,26 +32,6 @@ import (
 	"knative.dev/eventing/test/lib/resources"
 	"knative.dev/eventing/test/lib/resources/sender"
 )
-
-// SetupInfrastructureFunc sets up the infrastructure for running tracing tests. It returns the
-// expected trace as well as a string that is expected to be in the logger Pod's logs.
-type SetupInfrastructureFunc func(
-	t *testing.T,
-	channel *metav1.TypeMeta,
-	client *lib.Client,
-	loggerPodName string,
-	tc TracingTestCase,
-) (tracinghelper.TestSpanTree, cetest.EventMatcher)
-
-// TracingTestCase is the test case information for tracing tests.
-type TracingTestCase struct {
-	// IncomingTraceId controls whether the original request is sent to the Broker/Channel already
-	// has a trace ID associated with it by the sender.
-	IncomingTraceId bool
-	// Istio controls whether the Pods being created for the test (sender, transformer, logger,
-	// etc.) have Istio sidecars. It does not affect the Channel Pods.
-	Istio bool
-}
 
 // ChannelTracingTestHelperWithChannelTestRunner runs the Channel tracing tests for all Channels in
 // the ChannelTestRunner.
@@ -66,90 +41,8 @@ func ChannelTracingTestHelperWithChannelTestRunner(
 	setupClient lib.SetupClientOption,
 ) {
 	channelTestRunner.RunTests(t, lib.FeatureBasic, func(t *testing.T, channel metav1.TypeMeta) {
-		ChannelTracingTestHelper(t, channel, setupClient)
+		tracingTest(t, setupClient, setupChannelTracingWithReply, channel)
 	})
-}
-
-// ChannelTracingTestHelper runs the Channel tracing test using the given TypeMeta.
-func ChannelTracingTestHelper(t *testing.T, channel metav1.TypeMeta, setupClient lib.SetupClientOption) {
-	testCases := map[string]TracingTestCase{
-		"includes incoming trace id": {
-			IncomingTraceId: true,
-		},
-	}
-
-	for n, tc := range testCases {
-		t.Run(n, func(t *testing.T) {
-			tracingTest(t, setupClient, setupChannelTracingWithReply, channel, tc)
-		})
-	}
-}
-
-func tracingTest(
-	t *testing.T,
-	setupClient lib.SetupClientOption,
-	setupInfrastructure SetupInfrastructureFunc,
-	channel metav1.TypeMeta,
-	tc TracingTestCase,
-) {
-	const (
-		recordEventsPodName = "recordevents"
-	)
-
-	client := lib.Setup(t, true, setupClient)
-	defer lib.TearDown(client)
-
-	// Do NOT call zipkin.CleanupZipkinTracingSetup. That will be called exactly once in
-	// TestMain.
-	tracinghelper.Setup(t, client)
-
-	// Setup the test infrastructure
-	expectedTestSpan, eventMatcher := setupInfrastructure(t, &channel, client, recordEventsPodName, tc)
-
-	// Start the event info store and assert the event was received correctly
-	targetTracker, err := recordevents.NewEventInfoStore(client, recordEventsPodName)
-	if err != nil {
-		t.Fatalf("Pod tracker failed: %v", err)
-	}
-	defer targetTracker.Cleanup()
-	matches := targetTracker.AssertAtLeast(1, recordevents.MatchEvent(eventMatcher))
-
-	// Match the trace
-	traceID := getTraceIDHeader(t, matches)
-	trace, err := zipkin.JSONTracePred(traceID, 5*time.Minute, func(trace []model.SpanModel) bool {
-		tree, err := tracinghelper.GetTraceTree(trace)
-		if err != nil {
-			return false
-		}
-		// Do not pass t to prevent unnecessary log output.
-		return len(expectedTestSpan.MatchesSubtree(nil, tree)) > 0
-	})
-	if err != nil {
-		t.Logf("Unable to get trace %q: %v. Trace so far %+v", traceID, err, tracinghelper.PrettyPrintTrace(trace))
-		tree, err := tracinghelper.GetTraceTree(trace)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if len(expectedTestSpan.MatchesSubtree(t, tree)) == 0 {
-			t.Fatalf("No matching subtree. want: %v got: %v", expectedTestSpan, tree)
-		}
-	}
-}
-
-// getTraceIDHeader gets the TraceID from the passed in events.  It returns the header from the
-// first matching event, but registers a fatal error if more than one traceid header is seen
-// in that message.
-func getTraceIDHeader(t *testing.T, evInfos []recordevents.EventInfo) string {
-	for i := range evInfos {
-		if nil != evInfos[i].HTTPHeaders {
-			sc := trace.RemoteSpanContextFromContext(trace.DefaultHTTPPropagator().Extract(context.TODO(), http.Header(evInfos[i].HTTPHeaders)))
-			if sc.HasTraceID() {
-				return sc.TraceIDString()
-			}
-		}
-	}
-	t.Fatalf("FAIL: No traceid in %d messages: (%s)", len(evInfos), evInfos)
-	return ""
 }
 
 // setupChannelTracing is the general setup for TestChannelTracing. It creates the following:
@@ -163,7 +56,7 @@ func setupChannelTracingWithReply(
 	channel *metav1.TypeMeta,
 	client *lib.Client,
 	recordEventsPodName string,
-	tc TracingTestCase,
+	senderPublishTrace bool,
 ) (tracinghelper.TestSpanTree, cetest.EventMatcher) {
 	eventSource := "sender"
 	// Create the Channels.
@@ -208,17 +101,17 @@ func setupChannelTracingWithReply(
 	// Everything is setup to receive an event. Generate a CloudEvent.
 	senderName := "sender"
 	eventID := uuid.New().String()
-	event := ce2.NewEvent()
+	event := cloudevents.NewEvent()
 	event.SetID(eventID)
 	event.SetSource(senderName)
 	event.SetType(lib.DefaultEventType)
 	body := fmt.Sprintf(`{"msg":"TestChannelTracing %s"}`, eventID)
-	if err := event.SetData(ce2.ApplicationJSON, []byte(body)); err != nil {
+	if err := event.SetData(cloudevents.ApplicationJSON, []byte(body)); err != nil {
 		t.Fatalf("Cannot set the payload of the event: %s", err.Error())
 	}
 
 	// Send the CloudEvent (either with or without tracing inside the SendEvents Pod).
-	if tc.IncomingTraceId {
+	if senderPublishTrace {
 		client.SendEventToAddressable(senderName, channelName, channel, event, sender.EnableTracing())
 	} else {
 		client.SendEventToAddressable(senderName, channelName, channel, event)
@@ -317,7 +210,7 @@ func setupChannelTracingWithReply(
 		},
 	}
 
-	if tc.IncomingTraceId {
+	if senderPublishTrace {
 		expected = tracinghelper.TestSpanTree{
 			// 1. Sending pod sends event to Channel (only if the sending pod generates a span).
 			Span: tracinghelper.MatchHTTPSpanNoReply(
